+++ tests/unit/test_input_data.py
@@ -2,10 +2,17 @@ import logging
 import os
 
 import pytest
+import scmdata
+import scmdata.testing
 import xarray as xr
 
+from spaemis.config import InputTimeseries
 from spaemis.constants import TEST_DATA_DIR
-from spaemis.input_data import InputEmissionsDatabase, initialize_database
+from spaemis.input_data import (
+    InputEmissionsDatabase,
+    initialize_database,
+    load_timeseries,
+)
 
 TEST_INPUT_DATA = os.path.join(TEST_DATA_DIR, "input4MIPs")
 
@@ -69,3 +76,61 @@ def test_initialize_data(monkeypatch, paths):
 
     res = initialize_database()
     assert res.paths == list(paths)
+
+
+def test_load_timeseries():
+    res = load_timeseries(
+        [
+            InputTimeseries(name="test", path="emissions_country.csv", filters=[{}]),
+            InputTimeseries(
+                name="test2",
+                path="emissions_country.csv",
+                filters=[dict(variable="Emissions|H2|Transportation Sector")],
+            ),
+            InputTimeseries(
+                name="test3",
+                path="emissions_country.csv",
+                filters=[
+                    dict(variable="Emissions|H2|Industrial Sector"),
+                    dict(region="World", keep=False),
+                ],
+            ),
+        ],
+        root_dir=os.path.join(TEST_DATA_DIR, "config"),
+    )
+
+    exp = scmdata.ScmRun(os.path.join(TEST_DATA_DIR, "config", "emissions_country.csv"))
+
+    scmdata.testing.assert_scmdf_almost_equal(
+        res["test"], exp, check_ts_names=False, allow_unordered=True
+    )
+    scmdata.testing.assert_scmdf_almost_equal(
+        res["test2"],
+        exp.filter(variable="Emissions|H2|Transportation Sector"),
+        check_ts_names=False,
+        allow_unordered=True,
+    )
+
+    scmdata.testing.assert_scmdf_almost_equal(
+        res["test3"],
+        exp.filter(variable="Emissions|H2|Industrial Sector").filter(
+            region="World", keep=False
+        ),
+        check_ts_names=False,
+        allow_unordered=True,
+    )
+
+
+def test_load_timeseries_duplicate():
+    with pytest.raises(ValueError, match="Duplicate input timeseries found: test"):
+        load_timeseries(
+            [
+                InputTimeseries(
+                    name="test", path="emissions_country.csv", filters=[{}]
+                ),
+                InputTimeseries(
+                    name="test", path="emissions_country.csv", filters=[{}]
+                ),
+            ],
+            root_dir=os.path.join(TEST_DATA_DIR, "config"),
+        )